requirements.txt
appdirs==1.4.3
asn1crypto==0.24.0
astroid==2.0.4
bottle==0.12.13
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
coverage==4.5.1
cryptography==2.3.1
Django==2.1.3
django-appconf==1.0.2
django-autoslug==1.9.3
django-compressor==2.2
django-debug-toolbar==1.10.1
django-haystack==2.8.1
django-ipware==2.1.0
django-mdeditor==0.1.9
django-pagedown==1.0.6
django-pure-pagination==0.3.0
django-uuslug==1.1.8
idna==2.7
isort==4.3.4
jieba==0.39
jsonpickle==1.0
lazy-object-proxy==1.3.1
Markdown==3.0.1
mccabe==0.6.1
mistune==0.8.4
olefile==0.46
packaging==18.0
Pillow==5.3.0
pycparser==2.19
Pygments==2.2.0
pylint==2.1.1
PyMySQL==0.9.2
pyparsing==2.3.0
python-memcached==1.59
python-slugify==1.2.6
pytz==2018.7
raven==6.9.0
rcssmin==1.0.6
requests==2.20.0
rjsmin==1.0.12
six==1.11.0
sqlparse==0.2.4
Unidecode==1.0.22
urllib3==1.24.1
webencodings==0.5.1
WeRoBot==1.6.0
Whoosh==2.7.4
wrapt==1.10.11
xmltodict==0.11.0
